run: provide scheduled start time for operations
Modifies the OperationContext structure so that it includes the
scheduled start time operation (which can be earlier than the `now`
timestamp).

When operations are scheduled to run with a fixed rate, it might happen
that an operation takes longer than usual and, due to fixed concurrency,
some other operations might be delayed. Measuring the latency starting
from the scheduled start time, not the actual start time, leads to more
realistic latencies being reported (in short, it accounts for the
coordinated omission problem).