add sharded_stats
Adds a helper module "sharded_stats" which adds the ShardedStats
structure which is optimized for frequent writes and seldom reads. Each
thread keeps its generated statistics in its own shards so that the
threads do not need to synchronize between themselves when updating
stats. On read, all shards are collected and joined, clearing the stats
accumulated so far.