s-b/args: add host selection policy flag
Adds a flag which allows specifying the load balancing policy (called
host selection policy in gocql). Currently, "round-robin" and
"token-aware" options are implemented from the original scylla-bench.
The "host-pool" policy is not implemented as it is based on an external
Go library which is unavailable for Rust. Additionally, the "dc-aware"
policy is implemented which, although it was declared in the original
s-b, it was not implemented there.