s-b/operation/read: handle timeout from session::execute_iter
In newer versions of the scylla-rust-driver, `session::execute_iter`
waits for the first page to be fetched before returning the iterator.
Because of this change and the fact that the driver doesn't support
timeouts natively yet, we need to work around it by using
tokio::time::timeout on the future returned from
`session::execute_iter`.