s-b: return a non-zero exit code when incorrect flags are passed
Previously, the main function returned Ok() after failing to parse the
CLI flags. Ok() here translates to a zero exit code, which may
incorrectly be interpreted as a success when e.g. used from a script.

This commit makes sure that Err() is returned in the aforementioned case
so that a non-zero exit code is returned.