run: add actual_start_time to OperationContext
The existing `scheduled_start_time` field represents the time when an
operation was supposed to start executing, and can be used to measure
operation latency adjusted for coordinated omission. This commit
introduces `actual_start_time` field which is set to the actual start
time of an operation and can be used to measure unadjusted latency.