s-b/args: add hdr-latency-file flag
The -hdr-latency-file option allows to specify a file to which latency
data will be written in the HistogramLogWriter-compatible format.

Although the Rust library for handling histograms that we use does not
support writing in this format, the commit adds a module which supports
the format. Implementation was modeled after the hdrhistogram-go
library.