s-b/main: set default log level verbosity to "warn"
By default, the env_logger crate uses the "error" level as the default.
However, in benchmarks, it makes sense to print messages with "warn"
verbosity as they may point to problems that affect performance.

This commit increases the default verbosity level to "warn". The log
level can still be set using the RUST_LOG environment variable - see the
documentation of the env_logger crate.