run: get rid of a per-operation allocation
Currently, Rust does not support async in traits natively and a usually
the "async-trait" crate is used to make async methods work. The main
limitation of async-trait is that it requires boxing the future returned
from the async function - so, an allocation.

The `Operation::execute` method is async, so it incurs an allocation for
each operation executed in the benchmark. While the driver itself
performs multiple allocations when handling a single request, for some
reason the allocation related to boxing the future was prominent in
flamegraphs of CPU-bound workloads (it took about 10% of the time).

This commit refactors the Operation so that it is responsible for
running its own `execute` loop in `Operation::run`. In order to make
migration to the new pattern easy, a macro `make_runnable!` is
introduced which generates an implementation of `run` based on the
existing `execute` method. Meanwhile, `execute` is demoted to a
non-trait async method, which Rust supports without any boxing.

This isn't a perfect abstraction because we can't force `run`
implementations to use `make_runnable!`, but the performance gains are
worth it. I couldn't find a way how to make the abstraction more tight;
I tried introducing a generic method which takes an async callback, but
it's very hard (if possible at all) to spell out the types and lifetimes
properly when the async callback takes non-static arguments and borrows
them in the returned future.