run: don't report an error in task when the run is stopped
Previously, when the run was signalled to stop, if a task fails an
operation and notices that it should stop, it returned an error. Now, it
doesn't, and only returns an error in case either all global retries are
exhausted or an error occured more than `max_retries_per_op` times.

The new semantic makes more sense. For example, if the test is stopped
manually with 10 consecutive retries allowed, and there is a task which
performed 5 operations and notices that it should stop, the task
shouldn't return an error because it might have succeeded if it tried
one more time.