run: change max_retries_per_op to max_consecutive_errors_per_op
The `max_retries_per_op` was introduced with the sb's
`error-at-row-limit` option, however the semantics of that option was
misunderstood. This option defines the maximum number of errors that
can happen for a single operation _in a row_, not the number of
retries. This commit changes the semantics and the name of that option
to `max_consecutive_errors_per_op`, which better aligns with the
original implementation. The `test_retrying` function is also
accordingly adjusted.