c-s rng: add java_random crate
Since c-s is implemented in Java, most of the generators base on
`java.util.Random` RNG. We want to behave in the exact same way
c-s does - meaning, generate the same values that c-s does.

Note that c-s is (most of the times, specifically, when using `-pop seq` parameter)
deterministic. This helps to validate the values inserted to DB.
First, we run the `write` workload, generating some values and inserting them to DB.
Afterwards, we can validate inserted data with the `read` workload.
Since the data that is inserted, is generated in a deterministic way,
we can (given the seed for generating PK), reconstruct whole row from scratch
when validating the data with `read` workload.

That's why we introduce the `java_random` crate, which is simply
Rust implementation of `java.util.Random`.