c-s values_generator: return CqlValue
Before the commit, the `ValueGenerator` would return a value of the
type that was associated to the specific implementation.

During the `read` workload (one of the following commits), we validate the
data, since we assume that DB was populated by a `write` workload in a
deterministic way.

Note that `QueryResult`'s rows contain `Vec<Option<CqlValue>>`, thus,
returning a CqlValue from the generator is much more convenient, since
comparison is trivial (I mean, the only obstacle is an `Option`, which
denotes `NULL`, however we don't really care about nulls at this point,
since we don't generate them in the write routine).