c-s operation: row_generator
This commit introduces operation module, as well as operation::row_generator
module. `RowGenerator` will be responsible for workload generation for
multiple scenarios.

In the `write` scenario - the row will be simply inserted to DB.
In the `read` scenario - we will fetch the row from the DB (where key=generated_key),
and then, the rest of the row data will be validated.

The row generation logic:
- sample `partition_key` (call it `pk_seed`) seed from the distribution shared between tasks
- seed the `partition_key` value generator with `pk_seed`
- generate the `partition_key`
- compute `columns_seed` based fully on `partition_key`.
- for each `column_generator`:
    - seed with `columns_seed`
    - generate the value for given column

As we can see, the row contents fully depend on partition_key - and more precisely,
on the `partition_key_seed` sampled from the shared distribution.