tracing: use tracing-subscriber
Before the commits, we would use `env_logger` to consume the logs
emitted with `tracing` crate. Since we included `tracing` crate with `log`
feature, there should be no problems.

However, I noticed that when using `env_logger`, all the logs appearing
after the first `await` in the code are simply ignored - they are not
displayed (the same applies to the rust-driver logs, which uses `tracing`
crate as well - https://rust-driver.docs.scylladb.com/stable/logging/logging.html).

I'm not sure where does this issue come from, but it's rather recommended
to use `tracing-subscriber` in `tracing` environment.

That's why we replace the `env_logger` crate with `tracing-subscriber`.
After the changes, all logs are consumed, and displayed in the console.