c-s parser: refactor
Before this commit, the parser would only be responsible for input VALIDATION.
The parameters' handles would return the strings that were matched to the
given parameter (and validated). However, the ideal approach is so the
user can retrieve the parsed value instead of validated String. This commit
introduces the changes addressing this issue.

The `SimpleParam` is now generic over `Parsable` trait, which is responsible
for parsing strings to the specified type (`type Parsable::Parsed`).

Adjusted the rest of the code to this change, including fixups in
read/write command parsing.