c-s param: change default to Option<&str>
Unfortunately, in the `Population` option (introduced in the following commit),
there is one parameter that requires a default value which is being dynamically
constructed (not `&'static str`, but `String`).

This commit prepares the rest of the code exactly for this parameter, thus
removes the static lifetime bound for the `default` value.