c-s simplify access to settings
Before this commit, `CommandParams` would be an enum, which is:
- `Special` - if some special command was executed (for now, it's only `help` command)
- `ReadWriteParams` - if one of `read`/`write`/`counter_read`/`counter_write` command was provided

However, passing such enum to the user who wants to execute the runtime (`ReadWriteParams` case),
is not a good idea. The user would always need to unwrap the workload parameters
by pattern matching (and what to do if `Special` is matched? panic? this should never happen,
since we return prematurely if `Special` is returned). It sounds really tedious to do so.

This commit refactors the `CommandParams` to struct, which contains:
- `basic_params: ReadWriteParams` - these are the parameters which are common
  for all of the types of runtime commands (including even `mixed` and `user` commands).
- in the future, it should contain some optional additional parameters,
  specific to either `mixed` or `user` command. Something like:
  `additional_params: Option<Either<MixedParams, UserParams>>`.

Regarding the parsing logic, we replace the `CommandParams::Special` with
`None::<CommandParams>`, and adjust all of the usages.