c-s param: get rid of `ParamMatchResult`
Before this commit, `Param::try_match()` function would return `ParamMatchResult`
enum which holds:
- `Match` if parameter successfully matched the provided argument
- `NoMatch` if argument didn't match the parameter
- `anyhow::Err` if some error occured.

It doesn't make much sense, for a function which checks whether the argument
matches the parameter, to return an error in any case.
This commit changes `try_match` so now it returns simply `bool`. This way,
we can also get rid of `ParamMatchResult` enum at all.

The error handling is now delegated to `parse` function. As we can see,
most of the logic, such as:
- checking whether user specified some parameter twice (self.supplied_by_user check)
- checking whether argument matches the parameter (`arg.starts_with(self.prefix)`)
- removing the argument's prefix, to retrieve the value (`arg_value = &arg[self.prefix.len()..]`)
is exactly the same for both `Simple` and `Multi` params, and so, is being repeated.

This change makes us realize (that's why it's a good change!)
that most of the parsing logic can be extracted to some struct wrapping
the `Param` implementations. This is exactly what the next commit introduces.