c-s read: add stats recording
Extended the `ReadOperation` logic so it records operation result
in the `ShardedStats` structure. Extracted read + validation logic to
separate function `do_execute`. It allows us to easily obtain the result
of operation, and to record the result to `ShardedStats` structure.

Adjusted `main` logic so it passes the `ShardedStats` to `ReadOperationFactory`.