c-s parser: introduce TypedParam
Before this commit, a lot of boilerplate code (such as getters for `required` parameter's flag)
would be copied in implementation of multiple parameters types. For reference, see: https://github.com/scylladb/cql-stress/pull/40#discussion_r1267975131 (comment).

Most of the common logic (as well as shared fields) can be delegated to some
structure which would wrap the specific implementations.

In this commit we introduce a `TypedParam` structure which wraps the specific
implementation of `ParamImpl`.

For now, all of the `TypedParam`'s methods are simply delegated to `ParamImpl` implementation,
without any specific changes. I prefer to split the changes into multiple commits
so the reviewers can understand the changes step by step. Later, this,
and the following commits regarding the parameters refactor can be squashed before
merge.