c-s types: impl Parsable for ConnectionsPerShard
-mode option supports `connectionsPerHost` parameter
which lets the user define the number of connections
opened to each host.

In cql-stress we will also support `connectionsPerShard` parameter
which sets the number of connections per shard.

Note that in the following commits, the `shard-connection-count`
will be moved from `-node` option to `-mode` option and renamed to
`connectionsPerShard` to follow the naming conventions from
cassandra-stress.

We implement Parsable for `ConnectionsPerShard` struct which
returns `scylla::transport::session::PoolSize::PerShard`.