row_generator: add name to index value mapping
Until now, bound values of prepared statements had fixed position, since
the statements themselves were predefined by the tool.

With support of user profiles, users can define custom queries
in which the order of bound values can be different.

This is why, the row generator should contain an information about the
mapping: column name to bound variable index. Thanks to that, we will be
able to correctly bind the values from generated row to the statement
execution context.