operation: fix `recompute_seed` logic
Seed for non-pk column generators are computed based on pk value.
`recompute_seed` function is responsible for computing such seed.

However, there is a small error in this function. All of the values
should be interpreted as SIGNED integers.

When we cast a `u8` value to `i64`, the result will always be positive.
This is why, we firstly need to cast it to `i8`, to interpret is as a
signed value and allow negative values.

Note: Previously, the error was not recognized since `HexBlob` generator
generates positive values that can be contained in 7 bits. The overflow
was not present there.