c-s/col: c-s compatibility for -col n= parameter
cassandra-stress supports a distribution syntax for `-col n=` parameter.
However, it only allows the `FIXED(?)` distribution for cql mode.

Before this commit, cql-stress would accept a u64 value for
the value of this parameter.

To make our CLI compatible with c-s (so it can be easily replaced in SCT),
we will accept `FIXED(<u64>)` syntax as well.