CLI/user: add `ops` and `clustering` parameters
Semantics of parameters:
- `ops` -> defines a ratio with which the operation/queries will be
sampled. For example ops(foo=1,bar=2) means that the tool will
approximately execute 2 `bar` queries per 1 `foo` query. These queries
must be defined in the profile yaml.

- `clustering` -> a distribution. The tool will sample a value from
this distribution to decide how many times to execute
current operation/query (which was sampled based on `ops`). The value
will be treated as a counter. When counter reaches 0, new operation
is sampled, and so, the new value of counter is sampled as well.