mixed: create factory only if operation is defined
Each operation factory prepares a statement to be executed by its
corresponding operation. The preparation happens when we construct
the factory (i.e. <TypeOfOperation>Factory::new), so the created
factory object can hold a prepared statement and pass it to the
operation objects that it creates.

A mixed command encapsulates 4 other operations/operation factories:
- {Read/Write/CounterRead/CounterWrite}Operation
- {Read/Write/CounterRead/CounterWrite}OperationFactory

We don't always want to prepare 4 statements (one for each operation type).
We decided that `mixed` command will not prepare a schema - this means
that the tool will not create any keyspaces/tables. The corresponding
write workload should be run before it, to create and populate a table.
Assume that someone ran a `write` command - it created `keyspace1.standard1` table.
As of now, if someone now ran a `mixed ratio(write=1,read=1)` command, the
`Counter{Read/Write}OperationFactory` constructors would prepare the
statements operating on `keyspace1.counter1` table - even though,
such table does not exist and the user will not be using this table
anyway (provided only write and read operations in ratio param).
This commit prevents it from happening - we only invoke the factory
constructor if user defined a corresponding operation to be sampled.

Note for reviewers:
I truly hate this design, but I have no idea how to implement it in
a better way without the support of async trait objects in Rust.
As of now, each operation needs to be handled separately
(a separate field for each operation/operation factory).
If Rust supported async trait objects, we could obviously hold the
operations in a collection (i.e. as Box<dyn CassandraStressOperation>).
Unfortunately, we are forced to use an `Option<>` for these fields
and unwrap (or handle an error that is not possible to occur) in multiple places.